hotbar indexをshortにする

+++ Assets/Scripts/MainGame/UnityView.Interface/PlayerInput/IBlockPlaceEvent.cs
@@ -4,7 +4,7 @@ namespace MainGame.UnityView.Interface.PlayerInput
 {
     public interface IBlockPlaceEvent
     {
-        public delegate void OnBlockPlace(Vector2Int position,int hotBarIndex);
+        public delegate void OnBlockPlace(Vector2Int position,short hotBarIndex);
 
         public void Subscribe(OnBlockPlace onBlockPlace);
     }
